+++ src/cmd/trace/goroutinegen.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package trace
+package main
 
 import (
 	tracev2 "internal/trace/v2"
