+++ src/cmd/trace/testdata/generate.go
