+++ src/cmd/trace/testdata/go122.test
