+++ src/cmd/trace/testdata/mktests.go
