+++ src/cmd/trace/testdata/testprog/main.go
